Removed periodic redraw, seems unnecessary. May reconsider.
Reorganized things in an effort to power on the EPD as briefly as
possible. Only drawing functions are called while the EPD is on;
all of the updating stuff runs before the display is powered on,
and remaining housekeeping is done after it's powered off.

Voltage display is broken (0.00V) when on battery power, at least
upon startup (next updating pending here). Worked when USB was
connected; maybe the EPD needs to be powered on before we can
read this value. Will fix later.