Test for presence of 'weather' and 'main' keys in JSON response.
This may catch an OWM reply that doesn't count as an HTTP failure
or JSON deserialization error, but also doesn't contain the fields
that we expect.

I witnessed a bad weather update that I couldn't explain (fields
were erased from the display and not redrawn). This may be the
reason why. Next step would be to test if any of the specific
values are null.